CHANGE Move functions to root level of RxStorage for WebWorker support
