+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### X.X.X
 
+Changes:
+  - RxStorage: The non async functions `prepareQuery`, `getSortComparator` and `getQueryMatcher` have been moved out of `RxStorageInstance` into `RxStorage`. This was needed to have better WebWorker support. This will not affect you do not use a custom `RxStorage` implementation.
+
 Bugfixes:
   - Do not throw an error when database is destroyed while a GraphQL replication is running.
