+++ src/plugins/key-compression.ts
@@ -125,7 +125,7 @@ export const RxDBKeyCompressionPlugin: RxPlugin = {
     hooks: {
         /**
          * replace the keys of a query-obj with the compressed keys
-         * because the storage instance only know the compressed schema
+         * because the storage instance only knows the compressed schema
          * @return compressed queryJSON
          */
         prePrepareQuery(
