+++ src/plugins/lokijs/rx-storage-lokijs.ts
@@ -1,12 +1,19 @@
+import { DeterministicSortComparator, QueryMatcher } from 'event-reduce-js';
+import { newRxError } from '../../rx-error';
+import { getPrimaryFieldOfPrimaryKey } from '../../rx-schema';
+import lokijs from 'lokijs';
 import type {
     LokiDatabaseSettings,
     LokiSettings,
     LokiStorageInternals,
+    MangoQuery,
+    RxDocumentWriteData,
+    RxJsonSchema,
     RxKeyObjectStorageInstanceCreationParams,
     RxStorage,
     RxStorageInstanceCreationParams
 } from '../../types';
-import { flatClone, hash } from '../../util';
+import { firstPropertyNameOfObject, flatClone, hash } from '../../util';
 import {
     createLokiStorageInstance,
     RxStorageInstanceLoki
@@ -15,6 +22,7 @@ import {
     createLokiKeyObjectStorageInstance,
     RxStorageKeyObjectInstanceLoki
 } from './rx-storage-key-object-instance-loki';
+import { getLokiSortComparator } from './lokijs-helper';
 
 export class RxStorageLoki implements RxStorage<LokiStorageInternals, LokiSettings> {
     public name = 'lokijs';
@@ -25,6 +33,87 @@ export class RxStorageLoki implements RxStorage<LokiStorageInternals, LokiSettin
 
     hash(data: Buffer | Blob | string): Promise<string> {
         return Promise.resolve(hash(data));
+    }
+
+    prepareQuery<RxDocType>(
+        schema: RxJsonSchema<RxDocType>,
+        mutateableQuery: MangoQuery<RxDocType>
+    ) {
+        const primaryKey = getPrimaryFieldOfPrimaryKey(schema.primaryKey);
+        if (Object.keys(mutateableQuery.selector).length > 0) {
+            mutateableQuery.selector = {
+                $and: [
+                    {
+                        _deleted: false
+                    },
+                    mutateableQuery.selector
+                ]
+            };
+        } else {
+            mutateableQuery.selector = {
+                _deleted: false
+            };
+        }
+
+        /**
+         * To ensure a deterministic sorting,
+         * we have to ensure the primary key is always part
+         * of the sort query.
+         */
+        if (!mutateableQuery.sort) {
+            mutateableQuery.sort = [{ [primaryKey]: 'asc' }] as any;
+        } else {
+            const isPrimaryInSort = mutateableQuery.sort
+                .find(p => firstPropertyNameOfObject(p) === primaryKey);
+            if (!isPrimaryInSort) {
+                mutateableQuery.sort.push({ [primaryKey]: 'asc' } as any);
+            }
+        }
+
+        return mutateableQuery;
+    }
+
+
+    getSortComparator<RxDocType>(
+        schema: RxJsonSchema<RxDocType>,
+        query: MangoQuery<RxDocType>
+    ): DeterministicSortComparator<RxDocType> {
+        return getLokiSortComparator(schema, query);
+    }
+
+    /**
+     * Returns a function that determines if a document matches a query selector.
+     * It is important to have the exact same logix as lokijs uses, to be sure
+     * that the event-reduce algorithm works correct.
+     * But LokisJS does not export such a function, the query logic is deep inside of
+     * the Resultset prototype.
+     * Because I am lazy, I do not copy paste and maintain that code.
+     * Instead we create a fake Resultset and apply the prototype method Resultset.prototype.find(),
+     * same with Collection.
+     */
+    getQueryMatcher<RxDocType>(
+        schema: RxJsonSchema<RxDocType>,
+        query: MangoQuery<RxDocType>
+    ): QueryMatcher<RxDocumentWriteData<RxDocType>> {
+        const fun: QueryMatcher<RxDocumentWriteData<RxDocType>> = (doc: RxDocumentWriteData<RxDocType>) => {
+            const docWithResetDeleted = flatClone(doc);
+            docWithResetDeleted._deleted = !!docWithResetDeleted._deleted;
+
+            const fakeCollection = {
+                data: [docWithResetDeleted],
+                binaryIndices: {}
+            };
+            Object.setPrototypeOf(fakeCollection, (lokijs as any).Collection.prototype);
+            const fakeResultSet: any = {
+                collection: fakeCollection
+            };
+            Object.setPrototypeOf(fakeResultSet, (lokijs as any).Resultset.prototype);
+            fakeResultSet.find(query.selector, true);
+
+            const ret = fakeResultSet.filteredrows.length > 0;
+            return ret;
+        }
+        return fun;
     }
 
     async createStorageInstance<RxDocType>(
